Deprecate case2 (rhoc normalization)
Usual dimensionless variable in terms of central density is bad.
A monotonic parameter of BE sequence is the density contrast, rho_c/rho_e.
When finding maximum on P-R curve, we minimize the P in terms of this
monotonic parameter. When there is turbulence, we want to do the minimization
at the fixed sonic radius r_s. However, when the normalization is done
in terms of rho_c, the dimensionless sonic radius changes with rho_c/rho_e,
making it complicated to do the minimization. It is much more natural to
think of a fixed P_ext (and therefore fixed rho_e).