Deprecate the removal of the duplicated cores.
Sometimes, the tracked core is the t_coll core of another sink
particle. Then, the subsequent tracking would result in the
identical set of preimages of that other sink particle.

But this indicates that a fragmentation has occurred!
We do not want to entirely remove these cases, but rather
investigate these fragmentating cases (e.g., how often does it
occur). We will be able to identify the fragmenting cases by
finding multiple sink particles that share the same t_crit core.